Move SDK into own top-level pkg
